+++ ichnaea/api/locate/tests/test_internal.py
@@ -22,7 +22,10 @@ class TestCountrySource(BaseSourceTest):
         result = self.source.search(query)
         self.check_model_result(result, country)
         self.check_stats(counter=[
-            self.api_type + '.source.test.all.internal.low.hit'])
+            (self.api_type + '.source',
+                ['key:test', 'country:none', 'source:internal',
+                 'accuracy:low', 'status:hit']),
+        ])
 
     def test_ambiguous_mcc(self):
         cell = CellFactory.build(mcc=234)
@@ -30,7 +33,10 @@ class TestCountrySource(BaseSourceTest):
         result = self.source.search(query)
         self.check_model_result(result, None)
         self.check_stats(counter=[
-            self.api_type + '.source.test.all.internal.low.miss'])
+            (self.api_type + '.source',
+                ['key:test', 'country:none', 'source:internal',
+                 'accuracy:low', 'status:miss']),
+        ])
 
     def test_wifi(self):
         wifis = WifiFactory.build_batch(2)